Refactor hook for company progress indexing
